Add initial code on coding conventions
This adds a guide which documents the coding conventions we have noted down as a team. It is strongly encouraged that the team actively contribute to this guide as and when they think a new convention has become known.
Coding conventions help us maintain consistency across our SFD code bases, making it easier to read, understand and review each other's work. This shared set of conventions will help us onboard new team members and support the projects for long term maintainability.